move dynamic loading related functions to the loaders module
