+++ fract4d/c/fract4dc/common.h
@@ -1,0 +1,14 @@
+#ifndef __COMMON_H_INCLUDED__
+#define __COMMON_H_INCLUDED__
+
+    #define OBTYPE_MODULE "module"
+    #define OBTYPE_POINTFUNC "pf"
+    #define OBTYPE_CMAP "cmap"
+    #define OBTYPE_IMAGE "image"
+    #define OBTYPE_SITE "site"
+    #define OBTYPE_WORKER "worker"
+    #define OBTYPE_IMAGE_WRITER "image_writer"
+    #define OBTYPE_FFH "ffh"
+    #define OBTYPE_ARENA "arena"
+
+#endif
